+++ source/Programs/Compositor/SSAO/ssaox.frag
@@ -18,7 +18,7 @@ uniform mediump vec2 TexelSize0;
 
 //----------------------------------------------------------------------------------------------------------------------
 MAIN_PARAMETERS
-IN(highp vec2 vUV0, TEXCOORD0)
+IN(mediump vec2 vUV0, TEXCOORD0)
 
 
 MAIN_DECLARATION
